% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexcelbridge_controls.R
\name{clear_rexcelbridge_cache}
\alias{clear_rexcelbridge_cache}
\title{Clear the package cache of downloaded data files}
\usage{
clear_rexcelbridge_cache()
}
\value{
Invisibly returns \code{NULL}. A message is printed indicating which
directory was cleared.
}
\description{
Deletes the entire cache directory used by the \strong{rexcelbridge} package to store
downloaded data files. Useful if you need to force re-download of data,
or free up disk space.
}
\examples{
\dontrun{
# Remove all cached data files so they will be re-downloaded on next use
clear_rexcelbridge_cache()
}
}
\concept{helpers}
