% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexcelbridge-package.R
\docType{package}
\name{rexcelbridge-package}
\alias{rexcelbridge}
\alias{rexcelbridge-package}
\title{rexcelbridge: Bridge Excel Add-in Formulas and RTD Feeds into R (Windows)}
\description{
Vendor-neutral helpers to evaluate Excel formulas (including RTD-based add-ins like DTN ProphetX and Bloomberg) from R via COM automation. Supports single-cell and spilled-array results with a pluggable readiness predicate to handle transient "Wait" tokens from providers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/you/rexcelbridge}
  \item Report bugs at \url{https://github.com/you/rexcelbridge/issues}
}

}
\keyword{internal}
