% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{symbols_county_price}
\alias{symbols_county_price}
\title{Simulator Helper Datasets}
\format{
A data frame with 212 rows and 7 columns covering Inf--Inf.
}
\source{
ARPC using data from prophetX
}
\usage{
data(symbols_county_price)
}
\description{
A combined dataset for symbols_county_price
}
\keyword{datasets}
